requirements.txt
PySimpleGUI
openai
requests
gspread
google-auth
pytest
